chore: inline test_support
