+++ crates/amaru-ledger/src/state/volatile.rs
@@ -25,9 +25,6 @@ pub use series::VolatileSeries;
 mod view;
 pub use view::VolatileView;
 
-#[cfg(test)]
-pub(crate) mod test_support;
-
 pub trait VolatileStore {
     fn is_empty(&self) -> bool;
     fn len(&self) -> usize;
@@ -41,4 +38,79 @@ pub trait VolatileStore {
     fn rollback_to<'a>(&mut self, point: &'a Point) -> Result<(), &'a Point>;
     fn clear(&mut self);
     fn iter(&self) -> impl Iterator<Item = &AnchoredVolatileFragment>;
+}
+
+/// Shared test fixtures for the volatile keystone proptests, used by both `series` and `db`: a
+/// generator of UTxO-lifecycle windows and the brute-force oracle the maintained aggregate is
+/// checked against.
+#[cfg(test)]
+pub(crate) mod test_support {
+    use std::sync::Arc;
+
+    use amaru_kernel::{Hash, MemoizedTransactionOutput, TransactionInput};
+    use proptest::prelude::*;
+
+    use crate::{state::diff_set::DiffSet, tests::fake_output};
+
+    pub(crate) const VOLATILE_WINDOW: usize = 6;
+
+    prop_compose! {
+        /// A window of [`DiffSet`]s where each tagged UTxO has a unique lifecycle: produced once and
+        /// optionally consumed at a strictly later index. Mirrors UTxO uniqueness, so a newest-first
+        /// walk has a well-defined answer for every key.
+        pub(crate) fn unique_lifecycle_diffs()(
+            plan in prop::collection::btree_map(
+                0u8..16,
+                (0usize..VOLATILE_WINDOW, prop::option::of(0usize..VOLATILE_WINDOW)),
+                0..16,
+            )
+        ) -> Vec<DiffSet<TransactionInput, Arc<MemoizedTransactionOutput>>> {
+            let mut diffs: Vec<DiffSet<TransactionInput, Arc<MemoizedTransactionOutput>>> =
+                (0..VOLATILE_WINDOW).map(|_| DiffSet::default()).collect();
+
+            for (tag, (produced_at, consume_offset)) in plan {
+                diffs[produced_at].produce(test_input(tag), Arc::new(fixed_output()));
+                if let Some(offset) = consume_offset {
+                    let consumed_at = produced_at + 1 + offset;
+                    if consumed_at < VOLATILE_WINDOW {
+                        diffs[consumed_at].consume(test_input(tag));
+                    }
+                }
+            }
+
+            diffs
+        }
+    }
+
+    pub(crate) fn test_input(tag: u8) -> TransactionInput {
+        TransactionInput { transaction_id: Hash::new([tag; 32]), index: 0 }
+    }
+
+    pub(crate) fn fixed_output() -> MemoizedTransactionOutput {
+        fake_output("61bbe56449ba4ee08c471d69978e01db384d31e29133af4546e6057335")
+    }
+
+    /// Brute-force oracle: resolve `input` by walking `diffs` newest -> oldest. First consumed -> `None`,
+    /// first produce -> `Some`. The reference the maintained aggregate is checked against.
+    pub(crate) fn naive_resolve<'a>(
+        diffs: &'a [DiffSet<TransactionInput, Arc<MemoizedTransactionOutput>>],
+        input: &TransactionInput,
+    ) -> Option<&'a Arc<MemoizedTransactionOutput>> {
+        for diff in diffs.iter().rev() {
+            if diff.consumed.contains(input) {
+                return None;
+            }
+            if let Some(output) = diff.produced.get(input) {
+                return Some(output);
+            }
+        }
+        None
+    }
+
+    pub(crate) fn naive_has_consumed(
+        diffs: &[DiffSet<TransactionInput, Arc<MemoizedTransactionOutput>>],
+        input: &TransactionInput,
+    ) -> bool {
+        diffs.iter().any(|diff| diff.consumed.contains(input))
+    }
 }
